+++ packages/fighting-components/icon/src/icon.vue
@@ -10,11 +10,12 @@
 
 <script lang="ts" setup>
   import { Props, Emits } from './icon'
+  import type { onClickInterface } from '@fighting-design/fighting-type'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)
 
-  const onClick = (evt: Event): void => {
+  const onClick: onClickInterface = (evt: PointerEvent): void => {
     emit('click', evt)
   }
 </script>