+++ packages/fighting-type/auto.d.ts
@@ -1,0 +1,7 @@
+export interface onClickInterface {
+  (evt: PointerEvent): void
+}
+
+export interface ordinaryFunctionInterface {
+  (): void
+}