+++ packages/fighting-type/button.d.ts
@@ -1,7 +1,3 @@
-export interface onClickInterface {
-  (evt: Event): void
-}
-
 export type buttonSize = 'large' | 'middle' | 'small' | 'mini'
 
 export type buttonTarget = '_blank' | '_self' | '_parent' | '_top'