+++ packages/fighting-utils/install.ts
@@ -1,4 +1,7 @@
-import type { installInterface, mainVNodeInterface } from '@fighting-design/fighting-type'
+import type {
+  installInterface,
+  mainVNodeInterface
+} from '@fighting-design/fighting-type'
 import type { App } from 'vue'
 
 /**
@@ -6,7 +9,10 @@ import type { App } from 'vue'
  * @param main 组件实例
  * @param name 组件名
  */
-export const install: installInterface = (main: mainVNodeInterface, name: string): void => {
+export const install: installInterface = (
+  main: mainVNodeInterface,
+  name: string
+): void => {
   main.install = (app: App): void => {
     app.component(name, main)
   }